+++ lib/flop_phoenix/pagination.ex
@@ -60,55 +60,6 @@ defmodule Flop.Phoenix.Pagination do
     |> Misc.deep_merge(opts)
   end
 
-  def build_path_fun(_meta, nil), do: fn _ -> nil end
-
-  def build_path_fun(meta, path) do
-    &build_path_to_page(path, &1, build_query_params(meta))
-  end
-
-  defp build_query_params(meta) do
-    meta.flop
-    |> ensure_page_based_params()
-    |> Flop.Phoenix.to_query(backend: meta.backend, for: meta.schema)
-  end
-
-  defp build_path_to_page(path, page, query_params) do
-    Flop.Phoenix.build_path(path, maybe_put_page(query_params, page))
-  end
-
-  @doc """
-  Takes a `Flop` struct and ensures that the only pagination parameters set are
-  `:page` and `:page_size`. `:offset` and `:limit` are set to nil.
-
-  ## Examples
-
-      iex> flop = %Flop{limit: 2}
-      iex> ensure_page_based_params(flop)
-      %Flop{
-        limit: nil,
-        offset: nil,
-        page: nil,
-        page_size: 2
-      }
-  """
-  @spec ensure_page_based_params(Flop.t()) :: Flop.t()
-  def ensure_page_based_params(%Flop{} = flop) do
-    %{
-      flop
-      | after: nil,
-        before: nil,
-        first: nil,
-        last: nil,
-        limit: nil,
-        offset: nil,
-        page_size: flop.page_size || flop.limit,
-        page: flop.page
-    }
-  end
-
-  defp maybe_put_page(params, 1), do: Keyword.delete(params, :page)
-  defp maybe_put_page(params, page), do: Keyword.put(params, :page, page)
-
   def attrs_for_page_link(page, page, opts) do
     add_page_link_aria_label(opts[:current_link_attrs], page, opts)
   end
